fix:not plotting returns by q when missing sectors
if there are missing groups/sectors in the data, this get_level_values("group").unique() will return only existing values.
so, if have 10 groups/sectors but 3 are missing, it will return 7
then this code will create 7 subplots.
but, the code that draw that - does this: mean_ret_by_q.groupby(level='group')) which includes the missing groups.
it will result in only 7 subplots but 10 data subsets (3 will be empty) to plot.
so - it will plot empty subplots AND miss some existing data subsets.

I solve this by plotting all 10 groups (in the example case. could be any number)
and leaving empty subplots where data is missing.